+++ server/.gitignore
@@ -1,3 +1,3 @@
 /venv
-/__pycache__
+**/__pycache__/
 .vscode
